feat : step1 구현
[FileController]
- 리턴값을 프론트로 반환하지 않아도 되기 때문에, status 상태코드만 넘겨주도록 했습니다.

[Dto]
- dto는 데이터를 전달하기 위한 붋변객체여야하기 때문에 dto를 record타입으로 지정했습니다.

[FileService]
- 필요한 파일을 찾을 때, fileName과 id 중 무엇을 이용할지 고민했습니다.
- 프론트가 어떻게 ui를 구현하느냐에 따라 api명세서가 달라질 것 같다는 생각을 했습니다.
- 하나의 메소드안에 다 처리하기보다는, 개별 메소드로 기능을 분리하려고 했습니다.

[ServiceException - ErrorCode]
- 현재 step_1에서는 FILE 관련 에러만 발생하기 때문에, 'FILE_'로 prefix를 설정했습니다.
- 추후에 폴더, 기타 에러도 같은 규칙을 적용하면 어떨까 생각했습니다.

[ServiceException]
- 생성자, 메소드 오버로딩을 통해서 다형성을 가진 함수호출이 가능해졌습니다.

[테스트코드]
- mockito라이브러리와 각종 메소드에 대한 이해가 더 필요하다고 생각합니다.
- 실제 기능의 동작과정을 가상으로 똑같이 따라간다고 생각하니까 조금 더 수월해졌습니다.
- 파일자료타입에 대한 이해가 부족한 것도 어려움을 겪은 이유중에 하나라고 생각합니다.
